Put bookkeeper in its own file
